LayoutTests/fast/shadow-dom/direct-focus-pseudo-does-not-match-in-author-stylesheet-expected.txt

This tests that -webkit-direct-focus is not recognized in author stylesheets

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS defaultFocus.matches(":focus") is true
PASS defaultFocus.matches(":-webkit-direct-focus") threw exception SyntaxError: The string did not match the expected pattern..
PASS inputInShadowRoot.matches(":focus") is true
PASS inputInShadowRoot.matches(":-webkit-direct-focus") threw exception SyntaxError: The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
